more dom management moved over to user-management
