+++ src/js/elements/popup.js
@@ -1,4 +1,4 @@
-import { className, MAX_COUNT, NO_SEARCH_RESULTS_MSG } from '../consts.js';
+import { className, MAX_COUNT } from '../consts.js';
 import Element from './elements.js';
 import { addClass, show, hide, xssEscape } from '../utils.js';
 
@@ -100,31 +100,6 @@ class Popup extends Element {
 
         li.appendChild(div);
         return li;
-    }
-
-    createUserSearchEmptyResults() {
-        let errorContainer = this.createDiv();
-        this._setClass(errorContainer, [className.SEARCH_ERROR_CONTAINER]);
-        this.emptySearchResults = errorContainer;
-
-        let emptySearchResults = this.createDiv();
-        this._setClass(emptySearchResults, [className.NO_SEARCH_RESULTS]);
-        this._setContent(emptySearchResults, NO_SEARCH_RESULTS_MSG);
-
-        let errorImage = this.createDiv();
-        this._setClass(errorImage, [className.SEARCH_ERROR_IMG]);
-
-        errorContainer.appendChild(errorImage);
-        errorContainer.appendChild(emptySearchResults);
-        return errorContainer;
-    }
-
-    removeEmptySearchResults() {
-        if(this.emptySearchResults) {
-            let emptySearchResults = this.emptySearchResults;
-            emptySearchResults.parentNode.removeChild(emptySearchResults);
-        }
-        this.emptySearchResults = null;
     }
 
     _createInvitePopup() {
